+++ pkg/service/web_context.go
@@ -1,8 +1,6 @@
 package service
 
 import (
-	"fmt"
-	"math"
 	"time"
 
 	"github.com/exograd/eventline/pkg/eventline"
@@ -63,37 +61,5 @@ func (ctx *WebContext) formatDate(t time.Time, format eventline.DateFormat) (s s
 }
 
 func (ctx *WebContext) FormatDuration(d time.Duration) (s string) {
-	us := d.Microseconds()
-
-	switch {
-	case us == 0:
-		return fmt.Sprintf("%dns", d.Nanoseconds())
-
-	case us < 1_000:
-		return fmt.Sprintf("%dµs", us)
-
-	case us < 1_000_000:
-		return fmt.Sprintf("%.0fms", math.Ceil(float64(us)/1_000.0))
-
-	case us < 60_000_000:
-		return fmt.Sprintf("%.0fs", math.Ceil(float64(us)/1_000_000.0))
-
-	case us < 3_600_000_000:
-		m := us / 60_000_000
-		s := math.Ceil(float64(us%60_000_000) / 1_000_000.0)
-		if s > 0 {
-			return fmt.Sprintf("%dm%.0fs", m, s)
-		} else {
-			return fmt.Sprintf("%dm", m)
-		}
-
-	default:
-		h := us / 3_600_000_000
-		m := math.Ceil(float64(us%3_600_000_000) / 60_000_000.0)
-		if m > 0.0 {
-			return fmt.Sprintf("%dh%.0fm", h, m)
-		} else {
-			return fmt.Sprintf("%dh", h)
-		}
-	}
+	return utils.FormatDuration(d)
 }