+++ pkg/utils/time.go
@@ -2,6 +2,7 @@ package utils
 
 import (
 	"fmt"
+	"math"
 	"time"
 )
 
@@ -39,4 +40,40 @@ func FormatRelativeDate(t time.Time, now time.Time) (s string) {
 	}
 
 	return
+}
+
+func FormatDuration(d time.Duration) string {
+	us := d.Microseconds()
+
+	switch {
+	case us == 0:
+		return fmt.Sprintf("%dns", d.Nanoseconds())
+
+	case us < 1_000:
+		return fmt.Sprintf("%dµs", us)
+
+	case us < 1_000_000:
+		return fmt.Sprintf("%.0fms", math.Ceil(float64(us)/1_000.0))
+
+	case us < 60_000_000:
+		return fmt.Sprintf("%.0fs", math.Ceil(float64(us)/1_000_000.0))
+
+	case us < 3_600_000_000:
+		m := us / 60_000_000
+		s := math.Ceil(float64(us%60_000_000) / 1_000_000.0)
+		if s > 0 {
+			return fmt.Sprintf("%dm%.0fs", m, s)
+		} else {
+			return fmt.Sprintf("%dm", m)
+		}
+
+	default:
+		h := us / 3_600_000_000
+		m := math.Ceil(float64(us%3_600_000_000) / 60_000_000.0)
+		if m > 0.0 {
+			return fmt.Sprintf("%dh%.0fm", h, m)
+		} else {
+			return fmt.Sprintf("%dh", h)
+		}
+	}
 }